links for registration and session
edit initializers/devise.rb to turn on scoped views
run rails generate devise:views users
follow this guide to add links: http://goo.gl/36aX7

NOTE: it isn't technically necessary to use scoped views
or to generate the views. But it is common to do this and
makes it more clear how devise is working.